# src/main/java/com/tbp/crud/model/AuthenticationRequest.java
package com.tbp.crud.model;
import lombok.Data;

import java.io.Serializable;

@Data
public class AuthenticationRequest implements Serializable {

    private String username;
    private String password;

}